:recycle: refactor(user, auth ui): ui auth와 user 구분
