+++ src/main/java/ac/dnd/dodal/application/user/service/AuthKakaoLoginService.java
@@ -7,9 +7,9 @@ import ac.dnd.dodal.domain.user.enums.UserExceptionCode;
 import ac.dnd.dodal.domain.user.enums.UserRole;
 import ac.dnd.dodal.domain.user.exception.UserNotFoundException;
 import ac.dnd.dodal.domain.user.model.User;
-import ac.dnd.dodal.ui.user.response.JwtTokenDto;
-import ac.dnd.dodal.ui.user.response.KakaoUserInfoDto;
-import ac.dnd.dodal.ui.user.response.UserInfoResponseDto;
+import ac.dnd.dodal.ui.auth.response.JwtTokenDto;
+import ac.dnd.dodal.ui.auth.response.KakaoUserInfoDto;
+import ac.dnd.dodal.ui.auth.response.UserInfoResponseDto;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
