+++ src/main/java/ac/dnd/dodal/application/user/service/AuthSignUpService.java
@@ -6,9 +6,9 @@ import ac.dnd.dodal.application.user.usecase.UserQueryUseCase;
 import ac.dnd.dodal.core.security.util.JwtUtil;
 import ac.dnd.dodal.domain.user.enums.UserRole;
 import ac.dnd.dodal.domain.user.model.User;
-import ac.dnd.dodal.ui.user.request.OAuthUserInfoRequestDto;
-import ac.dnd.dodal.ui.user.response.JwtTokenDto;
-import ac.dnd.dodal.ui.user.response.UserInfoResponseDto;
+import ac.dnd.dodal.ui.auth.request.OAuthUserInfoRequestDto;
+import ac.dnd.dodal.ui.auth.response.JwtTokenDto;
+import ac.dnd.dodal.ui.auth.response.UserInfoResponseDto;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
