+++ src/main/java/ac/dnd/dodal/application/user/usecase/UserCommandUseCase.java
@@ -1,7 +1,7 @@
 package ac.dnd.dodal.application.user.usecase;
 
 import ac.dnd.dodal.domain.user.model.User;
-import ac.dnd.dodal.ui.user.request.OAuthUserInfoRequestDto;
+import ac.dnd.dodal.ui.auth.request.OAuthUserInfoRequestDto;
 
 public interface UserCommandUseCase {
     User createUserBySocialSignUp(OAuthUserInfoRequestDto authSignUpRequestDto);
