+++ src/main/java/ac/dnd/dodal/ui/auth/response/JwtTokenDto.java
@@ -1,7 +1,6 @@
-package ac.dnd.dodal.ui.user.response;
+package ac.dnd.dodal.ui.auth.response;
 
 import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
 
 
 public record JwtTokenDto(
