+++ src/test/java/ac/dnd/dodal/application/user/dto/UserCommandFixture.java
@@ -1,7 +1,7 @@
 package ac.dnd.dodal.application.user.dto;
 
 import ac.dnd.dodal.domain.user.UserFixture;
-import ac.dnd.dodal.ui.user.request.OAuthUserInfoRequestDto;
+import ac.dnd.dodal.ui.auth.request.OAuthUserInfoRequestDto;
 
 public class UserCommandFixture {
     private static final String APPLE_ID = "appleId";
