+++ src/test/java/ac/dnd/dodal/domain/user/UserFixture.java
@@ -2,7 +2,7 @@ package ac.dnd.dodal.domain.user;
 
 import ac.dnd.dodal.domain.user.enums.UserRole;
 import ac.dnd.dodal.domain.user.model.User;
-import ac.dnd.dodal.ui.user.request.OAuthUserInfoRequestDto;
+import ac.dnd.dodal.ui.auth.request.OAuthUserInfoRequestDto;
 
 public class UserFixture {
